populate data

+++ cmd/api/schema.go
@@ -22,7 +22,7 @@ type MovieGetInput struct {
 
 type MovieGetOutput struct {
 	Body struct {
-		ID        int64     `json:"ID" doc:"Unique integer ID for the movie"`
+		ID        uint     `json:"ID" doc:"Unique integer ID for the movie"`
 		CreatedAt time.Time `json:"createdAt" doc:"Timestamp for when the movie is added to our database"`
 		Title     string    `json:"title" doc:"Movie title"`
 		Year      int32     `json:"year" doc:"Movie release year"`
